refactor: seperate components and context
