+++ packages/assignment/src/components/ComplexForm.tsx
@@ -1,0 +1,90 @@
+// ComplexForm 컴포넌트
+import {renderLog} from "../utils.ts";
+import {useNotification} from "../context/notificationContext.tsx";
+import {useState} from "react";
+
+export const ComplexForm: React.FC = () => {
+    renderLog('ComplexForm rendered');
+    const {addNotification} = useNotification();
+    const [formData, setFormData] = useState({
+        name: '',
+        email: '',
+        age: 0,
+        preferences: [] as string[],
+    });
+
+    const handleSubmit = (e: React.FormEvent) => {
+        e.preventDefault();
+        addNotification('폼이 성공적으로 제출되었습니다');
+    };
+
+    const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        const {name, value} = e.target;
+        setFormData((prev) => ({
+            ...prev,
+            [name]: name === 'age' ? parseInt(value) || 0 : value,
+        }));
+    };
+
+    const handlePreferenceChange = (preference: string) => {
+        setFormData((prev) => ({
+            ...prev,
+            preferences: prev.preferences.includes(preference)
+                ? prev.preferences.filter((p) => p !== preference)
+                : [...prev.preferences, preference],
+        }));
+    };
+
+    return (
+        <div className='w-full md:w-1/2 md:pl-4'>
+            <div className='mt-8'>
+                <h2 className='text-2xl font-bold mb-4'>복잡한 폼</h2>
+                <form onSubmit={handleSubmit} className='space-y-4'>
+                    <input
+                        type='text'
+                        name='name'
+                        value={formData.name}
+                        onChange={handleInputChange}
+                        placeholder='이름'
+                        className='w-full p-2 border border-gray-300 rounded text-black'
+                    />
+                    <input
+                        type='email'
+                        name='email'
+                        value={formData.email}
+                        onChange={handleInputChange}
+                        placeholder='이메일'
+                        className='w-full p-2 border border-gray-300 rounded text-black'
+                    />
+                    <input
+                        type='number'
+                        name='age'
+                        value={formData.age}
+                        onChange={handleInputChange}
+                        placeholder='나이'
+                        className='w-full p-2 border border-gray-300 rounded text-black'
+                    />
+                    <div className='space-x-4'>
+                        {['독서', '운동', '음악', '여행'].map((pref) => (
+                            <label key={pref} className='inline-flex items-center'>
+                                <input
+                                    type='checkbox'
+                                    checked={formData.preferences.includes(pref)}
+                                    onChange={() => handlePreferenceChange(pref)}
+                                    className='form-checkbox h-5 w-5 text-blue-600'
+                                />
+                                <span className='ml-2'>{pref}</span>
+                            </label>
+                        ))}
+                    </div>
+                    <button
+                        type='submit'
+                        className='bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded'
+                    >
+                        제출
+                    </button>
+                </form>
+            </div>
+        </div>
+    );
+};
